move getUpgradesAvailable to separate file to make it mockable
